@@ -40,53 +40,56 @@
     "postinstall": "cd src/api/ && composer update && composer install --optimize-autoloader && cd ../../"
   },
   "dependencies": {
-    "@angular/animations": "^6.1.8",
-    "@angular/common": "^6.1.8",
-    "@angular/compiler": "^6.1.8",
-    "@angular/core": "^6.1.8",
-    "@angular/forms": "^6.1.8",
-    "@angular/http": "^6.1.8",
-    "@angular/platform-browser": "^6.1.8",
-    "@angular/platform-browser-dynamic": "^6.1.8",
-    "@angular/router": "^6.1.8",
-    "chartist": "^0.11.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "ajv": "^6.10.2",
+    "chartist": "^0.11.4",
     "chartist-plugin-tooltips": "^0.0.17",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^2.5.7",
+    "core-js": "^3.6.1",
     "dragula": "^3.7.2",
-    "highlight.js": "^9.12.0",
-    "marked": "^0.5.1",
-    "ng2-dragula": "^2.1.0",
-    "node-sass": "^4.9.3",
-    "rxjs": "^6.3.3",
+    "highlight.js": "^9.17.1",
+    "marked": "^0.8.0",
+    "ng2-dragula": "^2.1.1",
+    "node-sass": "^4.13.0",
+    "rxjs": "^6.5.4",
     "scss-base": "^1.4.0",
-    "zone.js": "^0.8.26"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.8.3",
-    "@angular/cli": "^6.2.3",
-    "@angular/compiler-cli": "^6.1.8",
-    "@angular/language-service": "^6.1.8",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~10.11.0",
-    "bourbon": "5.1.0",
-    "bourbon-neat": "1.9.1",
-    "codelyzer": "^4.4.4",
-    "jasmine": "^3.2.0",
-    "jasmine-core": "^3.2.1",
+    "@angular-devkit/build-angular": "^0.803.21",
+    "@angular/cli": "^8.3.21",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@types/chartist": "^0.9.46",
+    "@types/highlight.js": "^9.12.3",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/marked": "^0.7.2",
+    "@types/node": "^12.12.22",
+    "bourbon": "6.0.0",
+    "bourbon-neat": "4.0.0",
+    "codelyzer": "^5.2.1",
+    "jasmine": "^3.5.0",
+    "jasmine-core": "^3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.4",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^1.3.1",
-    "karma-phantomjs-launcher": "^1.0.4",
-    "nodemon": "^1.18.4",
-    "npm-run-all": "^4.1.3",
-    "protractor": "~5.4.1",
-    "ts-node": "~7.0.1",
-    "tslint": "~5.11.0",
-    "typescript": "^2.9.2"
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.5.1",
+    "nodemon": "^2.0.2",
+    "npm-run-all": "^4.1.5",
+    "protractor": "~5.4.2",
+    "puppeteer": "^2.0.0",
+    "ts-node": "^8.5.4",
+    "tslint": "~5.20.1",
+    "typescript": "^3.5.3"
   }
 }
