@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-component',
+  selector: 'tb-app-component',
   templateUrl: './app.component.html'
 })
 export class AppComponent {
