@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { BrowserModule, Title } from '@angular/platform-browser';
+import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { RouterModule } from '@angular/router';
@@ -10,7 +10,7 @@ import { ROUTES } from './app.routes';
 import { AppComponent } from './app.component';
 import { ApiInterceptor } from './app.api-http';
 
-import { Login } from './login/login.component';
+import { LoginComponent } from './login/login.component';
 import { BoardModule } from './board/board.module';
 import { DashboardModule } from './dashboard/dashboard.module';
 import { SettingsModule } from './settings/settings.module';
@@ -28,18 +28,19 @@ import { SharedModule } from './shared/shared.module';
     SharedModule,
     RouterModule.forRoot(ROUTES)
   ],
-  providers: [
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: ApiInterceptor,
-      multi: true
-    }
-  ],
+
+  providers: [{
+    provide: HTTP_INTERCEPTORS,
+    useClass: ApiInterceptor,
+    multi: true
+  }],
+
   declarations: [
     AppComponent,
-    Login
+    LoginComponent
   ],
-  bootstrap: [ AppComponent ]
+
+  bootstrap: [AppComponent]
 })
 export class AppModule { }
 
