@@ -4,9 +4,9 @@ import { Component, Input } from '@angular/core';
   selector: 'tb-calendar',
   templateUrl: './calendar.component.html'
 })
-export class Calendar {
-  private today: Date;
-  private tasks: Array<any>; // TODO: Use Task model when created
+export class CalendarComponent {
+  public today: Date;
+  public tasks: any[]; // TODO: Use Task model when created
 
   public dayLabels = [ 'Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat' ];
   public monthLabels = [
@@ -17,8 +17,9 @@ export class Calendar {
 
   public month: number;
   public year: number;
-  public calendarDays: Array<Array<string>>;
+  public calendarDays: string[][];
 
+  // tslint:disable-next-line
   @Input('board-id') boardId: number;
 
   constructor() {
@@ -49,7 +50,7 @@ export class Calendar {
     this.generateCalendar();
   }
 
-  public previousMonth(): void {
+  previousMonth(): void {
     this.month -= 1;
 
     // Months are zero-index in JS
@@ -61,7 +62,7 @@ export class Calendar {
     this.generateCalendar();
   }
 
-  public nextMonth(): void {
+  nextMonth(): void {
     this.month += 1;
 
     if (this.month > 11) {
@@ -72,21 +73,21 @@ export class Calendar {
     this.generateCalendar();
   }
 
-  private getColor(task: any) { // TODO: Use Task model
+  getColor(task: any) { // TODO: Use Task model
     return task.color;
   }
 
   private generateCalendar(): void {
-    let firstDate = new Date(this.year, this.month, 1),
-    startingDay = firstDate.getDay(),
-    monthLength = new Date(this.year, this.month + 1, 0).getDate(),
-    rows = Math.ceil((monthLength + startingDay) / 7),
-    day = 1;
+    const firstDate = new Date(this.year, this.month, 1);
+    const startingDay = firstDate.getDay();
+    const monthLength = new Date(this.year, this.month + 1, 0).getDate();
+    const rows = Math.ceil((monthLength + startingDay) / 7);
+    let day = 1;
 
     this.calendarDays = [];
 
     for (let i = 0; i < rows; ++i) {
-      let weekDays: Array<string> = [];
+      const weekDays: Array<string> = [];
 
       for (let j = 0; j < 7; ++j) {
         if (day <= monthLength && (j >= startingDay || i > 0)) {
