@@ -5,15 +5,20 @@ import * as Chartist from 'chartist';
   selector: 'tb-charts',
   templateUrl: './charts.component.html'
 })
-export class Charts implements AfterViewInit {
+export class ChartsComponent implements AfterViewInit {
   public percentages: Array<number>;
   public data: Array<number>;
   public words: Array<string>;
 
+  // tslint:disable-next-line
   @Input('chart-name') chartName = '';
+  // tslint:disable-next-line
   @Input('chart-type') chartType = 'pie';
+  // tslint:disable-next-line
   @Input('series') series = '';
+  // tslint:disable-next-line
   @Input('labels') labels = '';
+  // tslint:disable-next-line
   @Input('table-head') tableHead = '';
 
   constructor() {
@@ -34,50 +39,54 @@ export class Charts implements AfterViewInit {
   }
 
   private createPieChart() {
-    let data = {
+    const data = {
       series: this.data,
       labels: this.words
-    },
-      options = {
-        donut: true,
-        donutWidth: 100,
-        labelInterpolationFnc: (label: string, index: number) => {
-          let value = this.data[index],
-            percent = Math.round(value /
-              this.data.reduce(Chartist.sum) * 100);
+    };
+    const options = {
+      donut: true,
+      donutWidth: 100,
+      labelInterpolationFnc: (label: string, index: number) => {
+        const value = this.data[index];
+        const percent = Math.round(value /
+          this.data.reduce(Chartist.sum) * 100);
 
-          if (this.percentages.length < this.data.length) {
-            this.percentages.push(percent);
-          }
-
-          return label + ' ' + percent + '%';
+        if (this.percentages.length < this.data.length) {
+          this.percentages.push(percent);
         }
-      },
-      pie = new Chartist.Pie('#' + this.chartName, data, options);
+
+        return label + ' ' + percent + '%';
+      }
+    };
+
+    // tslint:disable-next-line
+    new Chartist.Pie('#' + this.chartName, data, options);
   }
 
   private createLineChart() {
-    let data = {
+    const data = {
       series: [this.data],
       labels: this.words
-    },
-      options = {
-        axisY: {
-          onlyInteger: true
-        },
-        low: 0,
-        height: '300px'// ,
-        // plugins: [ Chartist.plugins.tooltip({
-        //     transformTooltipTextFnc: (value) => {
-        //         return value + ' points remaining'
-        //     }
-        // }) ]
+    };
+    const options = {
+      axisY: {
+        onlyInteger: true
       },
-      line = new Chartist.Line('#' + this.chartName, data, options);
+      low: 0,
+      height: '300px'// ,
+      // plugins: [ Chartist.plugins.tooltip({
+      //     transformTooltipTextFnc: (value) => {
+      //         return value + ' points remaining'
+      //     }
+      // }) ]
+    };
+
+    // tslint:disable-next-line
+    new Chartist.Line('#' + this.chartName, data, options);
   }
 
   private convertToNumberArray(arr: Array<string>): Array<number> {
-    let nums: Array<number> = [];
+    const nums: Array<number> = [];
 
     for (let i = 0, len = arr.length; i < len; ++i) {
       nums.push(Number(arr[i]));
