@@ -3,9 +3,9 @@ import { CommonModule } from '@angular/common';
 
 import { SharedModule } from '../shared/shared.module';
 
-import { Dashboard } from './dashboard.component';
-import { Calendar } from './calendar/calendar.component';
-import { Charts } from './charts/charts.component';
+import { DashboardComponent } from './dashboard.component';
+import { CalendarComponent } from './calendar/calendar.component';
+import { ChartsComponent } from './charts/charts.component';
 
 @NgModule({
   imports: [
@@ -13,14 +13,14 @@ import { Charts } from './charts/charts.component';
     SharedModule
   ],
   declarations: [
-    Dashboard,
-    Calendar,
-    Charts
+    DashboardComponent,
+    CalendarComponent,
+    ChartsComponent
   ],
   exports: [
-    Dashboard,
-    Calendar,
-    Charts
+    DashboardComponent,
+    CalendarComponent,
+    ChartsComponent
   ]
 })
 export class DashboardModule { }
