@@ -4,14 +4,7 @@ import { HttpClient } from '@angular/common/http';
 import { Observable, of } from 'rxjs';
 import { map, catchError } from 'rxjs/operators';
 
-import {
-  ApiResponse,
-  Board,
-  Column,
-  Category,
-  IssueTracker,
-  User
-} from '../../shared/models';
+import { ApiResponse } from '../../shared/models';
 import { BoardData } from './board-data.model';
 
 @Injectable()
@@ -22,24 +15,24 @@ export class BoardAdminService {
   addBoard(board: BoardData): Observable<ApiResponse> {
     return this.http.post('api/boards', board)
     .pipe(
-      map((response: ApiResponse) => { return response; }),
-      catchError((err) => { return of(<ApiResponse>err.error); })
+      map((response: ApiResponse) =>  response),
+      catchError((err) =>  of(err.error as ApiResponse))
     );
   }
 
   editBoard(board: BoardData): Observable<ApiResponse> {
     return this.http.post('api/boards/' + board.id, board)
     .pipe(
-      map((response: ApiResponse) => { return response; }),
-      catchError((err) => { return of(<ApiResponse>err.error); })
+      map((response: ApiResponse) =>  response),
+      catchError((err) =>  of(err.error as ApiResponse))
     );
   }
 
   removeBoard(boardId: number): Observable<ApiResponse> {
     return this.http.delete('api/boards/' + boardId)
     .pipe(
-      map((response: ApiResponse) => { return response; }),
-      catchError((err) => { return of(<ApiResponse>err.error); })
+      map((response: ApiResponse) =>  response),
+      catchError((err) =>  of(err.error as ApiResponse))
     );
   }
 
