@@ -2,20 +2,22 @@ import { Component, OnDestroy } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 
 import { StringsService } from '../shared/services';
+import { SettingsService } from './settings.service';
 
 @Component({
   selector: 'tb-settings',
   templateUrl: './settings.component.html'
 })
-export class Settings implements OnDestroy {
+export class SettingsComponent implements OnDestroy {
   public strings: any;
   private subs: Array<any>;
 
-  constructor(private stringsService: StringsService,
-              private title: Title) {
+  constructor(public stringsService: StringsService,
+              public settings: SettingsService,
+              public title: Title) {
     this.subs = [];
 
-    let sub = stringsService.stringsChanged.subscribe(newStrings => {
+    const sub = stringsService.stringsChanged.subscribe(newStrings => {
       this.strings = newStrings;
       title.setTitle('TaskBoard - ' + this.strings['settings']); // tslint:disable-line
     });
