@@ -2,8 +2,8 @@ import { AppModule } from '../../src/app/app.module';
 
 describe('Module', () => {
   it('provides an AppModule class', () => {
-    let appModule = new AppModule();
+    const appModule = new AppModule();
     expect(appModule).toEqual(jasmine.any(Object));
   });
-})
+});
 
