@@ -1,4 +1,4 @@
-import { TestBed, ComponentFixture } from '@angular/core/testing'
+import { TestBed, ComponentFixture } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { FormsModule } from '@angular/forms';
@@ -8,8 +8,7 @@ import { Location } from '@angular/common';
 import { Router } from '@angular/router';
 
 import { BehaviorSubject } from 'rxjs';
-import { DragulaService } from 'ng2-dragula/ng2-dragula';
-import { DragulaModule } from 'ng2-dragula/ng2-dragula';
+import { DragulaService, DragulaModule } from 'ng2-dragula';
 
 import {
   AuthService,
@@ -18,19 +17,19 @@ import {
   ContextMenuService,
   NotificationsService
 } from '../../../src/app/shared/services';
-import { Login } from '../../../src/app/login/login.component';
+import { LoginComponent } from '../../../src/app/login/login.component';
 import { SettingsModule } from '../../../src/app/settings/settings.module';
 import { SharedModule } from '../../../src/app/shared/shared.module';
 import { DashboardModule } from '../../../src/app/dashboard/dashboard.module';
-import { BoardDisplay } from '../../../src/app/board/board.component';
+import { BoardDisplayComponent } from '../../../src/app/board/board.component';
 import { BoardService } from '../../../src/app/board/board.service';
-import { ColumnDisplay } from '../../../src/app/board/column/column.component';
-import { TaskDisplay } from '../../../src/app/board/task/task.component';
+import { ColumnDisplayComponent } from '../../../src/app/board/column/column.component';
+import { TaskDisplayComponent } from '../../../src/app/board/task/task.component';
 import { RouterMock, BoardServiceMock, DragulaMock } from '../mocks';
 
 describe('BoardDisplay', () => {
-  let component: BoardDisplay,
-    fixture: ComponentFixture<BoardDisplay>;
+  let component: BoardDisplayComponent;
+  let fixture: ComponentFixture<BoardDisplayComponent>;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -44,10 +43,10 @@ describe('BoardDisplay', () => {
         DashboardModule
       ],
       declarations: [
-        Login,
-        BoardDisplay,
-        ColumnDisplay,
-        TaskDisplay
+        LoginComponent,
+        BoardDisplayComponent,
+        ColumnDisplayComponent,
+        TaskDisplayComponent
       ],
       providers: [
         Title,
@@ -71,7 +70,7 @@ describe('BoardDisplay', () => {
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(BoardDisplay);
+    fixture = TestBed.createComponent(BoardDisplayComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
@@ -85,7 +84,7 @@ describe('BoardDisplay', () => {
     component.ngOnInit();
     expect(component.boardNavId).toEqual(0);
 
-    component.activeUser = <any>{ default_board_id: 2 };
+    component.activeUser = { default_board_id: 2 } as any;
     component.ngOnInit();
 
     const location = TestBed.get(Location);
@@ -94,14 +93,14 @@ describe('BoardDisplay', () => {
   });
 
   it('sets up drag and drop during ngAfterContentInit', () => {
-    component.activeBoard = <any>{ columns: [
+    component.activeBoard = { columns: [
       { id: 1, tasks: [{}] }
-    ] };
+    ] } as any;
     component.ngAfterContentInit();
 
-    expect((<any>component.dragula).opts.moves).toEqual(jasmine.any(Function));
+    expect((component.dragula as any).opts.moves).toEqual(jasmine.any(Function));
 
-    const test = (<any>component.dragula).opts.moves(null, null, {
+    const test = (component.dragula as any).opts.moves(null, null, {
       classList: { contains: () => false }
     });
     expect(test).toEqual(false);
@@ -119,9 +118,9 @@ describe('BoardDisplay', () => {
   });
 
   it('has a function to toggle filtered tasks', () => {
-    component.activeBoard = <any>{ columns: [
+    component.activeBoard = { columns: [
       { tasks: [{ hidefiltered: true }] }
-    ] };
+    ] } as any;
 
     component.toggleFiltered();
     expect(component.activeBoard.columns[0].tasks[0].hideFiltered).toEqual(false);
@@ -130,20 +129,20 @@ describe('BoardDisplay', () => {
   it('has a function to filter tasks', () => {
     component.userFilter = -1;
 
-    component.activeBoard = <any>{
+    component.activeBoard = {
       columns: [{
         tasks: [{
           assignees: []
         }]
       }]
-    };
+    } as any;
 
     const task = component.activeBoard.columns[0].tasks[0];
 
     component.filterTasks();
     expect(task.filtered).toEqual(false);
 
-    task.assignees = <any>[{ id: 1 }];
+    task.assignees = [{ id: 1 }] as any;
 
     component.filterTasks();
     expect(task.filtered).toEqual(true);
@@ -158,7 +157,7 @@ describe('BoardDisplay', () => {
     component.filterTasks();
     expect(task.filtered).toEqual(false);
 
-    task.categories = <any>[{ id: 1 }];
+    task.categories = [{ id: 1 }] as any;
 
     component.filterTasks();
     expect(task.filtered).toEqual(true);
@@ -170,13 +169,13 @@ describe('BoardDisplay', () => {
 
   it('updates the active board from a service', () => {
     component.boardService.updateActiveBoard(null);
-    component.boardService.updateActiveBoard(<any>{});
+    component.boardService.updateActiveBoard({} as any);
 
     expect(component.activeBoard).toEqual(jasmine.any(Object));
   });
 
   it('updates the active user from a service', () => {
-    component.auth.updateUser(<any>{ security_level: 1 });
+    component.auth.updateUser({ security_level: 1 } as any);
 
     expect(component.activeUser).toEqual(jasmine.any(Object));
   });
