@@ -1,14 +1,14 @@
-import { TestBed, ComponentFixture } from '@angular/core/testing'
+import { TestBed, ComponentFixture } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { Title } from '@angular/platform-browser';
 
-import { Dashboard } from '../../../src/app/dashboard/dashboard.component';
+import { DashboardComponent } from '../../../src/app/dashboard/dashboard.component';
 import { DashboardModule } from '../../../src/app/dashboard/dashboard.module';
 
 describe('DashBoard', () => {
-  let component: Dashboard,
-    fixture: ComponentFixture<Dashboard>;
+  let component: DashboardComponent;
+  let fixture: ComponentFixture<DashboardComponent>;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -24,7 +24,7 @@ describe('DashBoard', () => {
   });
 
   it('can be constructed', () => {
-    fixture = TestBed.createComponent(Dashboard);
+    fixture = TestBed.createComponent(DashboardComponent);
     component = fixture.componentInstance;
 
     expect(component).toBeTruthy();
