@@ -1,4 +1,4 @@
-import { TestBed } from '@angular/core/testing'
+import { TestBed } from '@angular/core/testing';
 import { Location } from '@angular/common';
 
 import { BehaviorSubject } from 'rxjs';
@@ -7,35 +7,46 @@ import { User, Board } from '../../src/app/shared/models';
 
 export class RouterMock {
   public url = {
-    indexOf: str => TestBed.get(Location).path().indexOf(str)
-  }
+    indexOf: (str: string) => TestBed.get(Location).path().indexOf(str)
+  };
 
-  navigate(arr) {
+  navigate(arr: any[]) {
     TestBed.get(Location).go(arr[0]);
   }
 }
 
 export class DragulaMock {
-  public opts;
-  public dropModel = new BehaviorSubject([
-    {},
-    { id: '1' },
-    { parentNode: { id: '1' } },
-    { parentNode: { id: '1' } }
-  ]);
+  public opts: any;
   public dragend = {
-    subscribe: (fn) => { fn(); }
+    subscribe: (fn: any) => { fn(); }
   };
 
-  find () {
+  find() {
     return { drake: {
       containers: []
     } };
   }
 
-  destroy () {}
+  dropModel() {
+    return {
+      subscribe: (fn: any) => {
+        fn([
+          {},
+          { id: '1' },
+          { parentNode: { id: '1' } },
+          { parentNode: { id: '1' } }
+        ]);
+      }
+    };
+  }
 
-  setOptions (name, opts) {
+  destroy() {}
+
+  createGroup(_: string, opts: any) {
+    this.opts = opts;
+  }
+
+  setOptions(_: any, opts: any) {
     this.opts = opts;
   }
 }
@@ -44,17 +55,17 @@ export class BoardServiceMock {
   public activeBoardChanged =
     new BehaviorSubject({ id: 0, name: 'Test', columns: [] });
 
-  getBoards () {
+  getBoards() {
     return new BehaviorSubject({
       data: [{}, [{ id: 1, name: 'Test', is_active: '1' }]]
     });
   }
 
-  updateActiveBoard (board) {
+  updateActiveBoard(board: any) {
     this.activeBoardChanged.next(board);
   }
 
-  updateColumn (col) {
+  updateColumn(_: any) {
     return new BehaviorSubject({});
   }
 }
@@ -66,21 +77,21 @@ export class SettingsServiceMock {
     { columns: [{ position: 3 }, { position: 2 }] }
   ]);
 
-  updateBoards () { }
+  updateBoards() { }
 
-  updateActions () { }
+  updateActions() { }
 
-  updateUsers () { }
+  updateUsers() { }
 
-  getUsers () {
+  getUsers() {
     return new BehaviorSubject({ data: [{}, [new User()]] });
   }
 
-  getBoards () {
+  getBoards() {
     return new BehaviorSubject({ data: [{}, [new Board()]] });
   }
 
-  getActions () {
+  getActions() {
     return new BehaviorSubject({ data: [{}, []] });
   }
 }
@@ -96,14 +107,14 @@ export class AuthServiceMock {
     language: 'en'
   };
 
-  authenticate () { return new BehaviorSubject(true); }
-  updateUser () { }
+  authenticate() { return new BehaviorSubject(true); }
+  updateUser() { }
 }
 
 export class NotificationsServiceMock {
   public noteAdded = new BehaviorSubject({});
 
-  addNote (note) {
+  addNote(note: any) {
     this.noteAdded.next(note);
   }
 }
